backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.3.5
pydantic==2.5.0
pydantic-settings==2.1.0
bcrypt==4.1.1
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
pandas==2.1.3
openpyxl==3.11.0
